finished setup for themes
