+++ src/styles/index.js
@@ -1,0 +1,1 @@
+export { textStyles } from "./textStyles/textStyles";
