+++ src/styles/textStyles/textStyles.js
@@ -1,0 +1,41 @@
+const textStyles = {
+  heading1: {
+    fontFamily: require('../../theme/fonts').type.montserratBold.name,
+    fontSize: require('../../theme/fonts').size.xxx_large,
+  },
+  heading2: {
+    fontFamily: require('../../theme/fonts').type.montserratBold.name,
+    fontSize: require('../../theme/fonts').size.xx_large,
+  },
+  heading3: {
+    fontFamily: require('../../theme/fonts').type.montserratMedium.name,
+    fontSize: require('../../theme/fonts').size.x_large,
+  },
+  heading4: {
+    fontFamily: require('../../theme/fonts').type.montserratSemiBold.name,
+    fontSize: require('../../theme/fonts').size.large,
+  },
+  //* here is the commons
+  subtitle: {
+    fontFamily: require('../../theme/fonts').type.montserratMedium.name,
+    fontSize: require('../../theme/fonts').size.large,
+  },
+  label: {
+    fontFamily: require('../../theme/fonts').type.montserratMedium.name,
+    fontSize: require('../../theme/fonts').size.large,
+  },
+  body1: {
+    fontFamily: require('../../theme/fonts').type.montserratRegular.name,
+    fontSize: require('../../theme/fonts').size.large,
+  },
+  body2: {
+    fontFamily: require('../../theme/fonts').type.montserratRegular.name,
+    fontSize: require('../../theme/fonts').size.medium,
+  },
+  caption: {
+    fontFamily: require('../../theme/fonts').type.montserratRegular.name,
+    fontSize: require('../../theme/fonts').size.small,
+  },
+}
+
+export { textStyles };
