--- src/utils/sizes.js
@@ -1,12 +1,0 @@
-const fontSize = {
-  xxx_large: 40,
-  xx_large: 32,
-  xxxx_large: 30,
-  x_large: 24,
-  large: 20,
-  medium: 16,
-  small: 14,
-};
-
-
-export { fontSize };
