--- src/utils/styles.js
@@ -1,41 +1,0 @@
-const textStyles = {
-  heading1: {
-    fontFamily: require('./fonts').fontFamily.montserratBold.name,
-    fontSize: require('./sizes').fontSize.xxx_large,
-  },
-  heading2: {
-    fontFamily: require('./fonts').fontFamily.montserratBold.name,
-    fontSize: require('./sizes').fontSize.xx_large,
-  },
-  heading3: {
-    fontFamily: require('./fonts').fontFamily.montserratMedium.name,
-    fontSize: require('./sizes').fontSize.x_large,
-  },
-  heading4: {
-    fontFamily: require('./fonts').fontFamily.montserratSemiBold.name,
-    fontSize: require('./sizes').fontSize.large,
-  },
-  //* here is the commons
-  subtitle: {
-    fontFamily: require('./fonts').fontFamily.montserratMedium.name,
-    fontSize: require('./sizes').fontSize.large,
-  },
-  label: {
-    fontFamily: require('./fonts').fontFamily.montserratMedium.name,
-    fontSize: require('./sizes').fontSize.large,
-  },
-  body1: {
-    fontFamily: require('./fonts').fontFamily.montserratRegular.name,
-    fontSize: require('./sizes').fontSize.large,
-  },
-  body2: {
-    fontFamily: require('./fonts').fontFamily.montserratRegular.name,
-    fontSize: require('./sizes').fontSize.medium,
-  },
-  caption: {
-    fontFamily: require('./fonts').fontFamily.montserratRegular.name,
-    fontSize: require('./sizes').fontSize.small,
-  },
-}
-
-export { textStyles };
